Adding error checking for the assembly. In particular we now check that the k-mers are within the limit of the velvet/oases and remove them if not. We now only delete assembly directories when everything went ok. All oases logs are kept. An error is returned if something went wrong.

+++ assemble.sh
@@ -2,7 +2,7 @@
 
 base=$1
 output=$2
-Ks=`echo $3 | sed s'/,/ /g'` #convert comma seperated to space separated
+Kall=`echo $3 | sed s'/,/ /g'` #convert comma seperated to space separated
 Kmerge=$4
 transLength=$5
 inputs="../../$6"
@@ -10,8 +10,22 @@ if [ $# -eq 7 ] ; then
    inputs="$inputs ../../$7"
 fi
 
+# Lets start by checking that the input k-mer lengths are supported
+Kmin=`velveth | grep "MAXKMERLENGTH" | cut -d"=" -f2`
+Ko=`oases | grep "MAXKMERLENGTH" | cut -d"=" -f2`
+if (( $Ko < $Kmin )) ; then Kmin=$Ko ; fi
+for k in $Kall ; do
+    if (( $k <= $Kmin )) ; then Ks="$Ks $k"
+    else 
+       echo "k=$k is above the minimum k-mer of your velvet/oases (${Kmin})."
+       echo "Lets keep going with out it..."
+    fi
+done
+
+#now start the real stuff
 mkdir ${base}/oases ; cd ${base}/oases
 for k in $Ks ; do
+    echo "Running Assembly for k="$k
     echo "directory_${k}" >> log_${k}
     echo "running velveth" >> log_${k}
     velveth directory_${k} $k -fastq -separate $inputs >> log_${k}
@@ -22,14 +36,32 @@ for k in $Ks ; do
     echo "done" >> log_${k}
 done
 
-echo "running velveth to merge with Kmerge = $Kmerge " >> log
-velveth mergedAssembly $Kmerge -long directory*/transcripts.fa >> log
-echo "running velvetg" >> log
-velvetg mergedAssembly -read_trkg yes -conserveLong yes >> log
-echo "running oases" >> log
-oases mergedAssembly -merge yes -min_trans_lgth $transLength >> log
-echo "done" >> log
+echo "running velveth to merge with Kmerge = $Kmerge " > log_merge
+velveth mergedAssembly $Kmerge -long directory*/transcripts.fa >> log_merge
+echo "running velvetg" >> log_merge
+velvetg mergedAssembly -read_trkg yes -conserveLong yes >> log_merge
+echo "running oases" >> log_merge
+oases mergedAssembly -merge yes -min_trans_lgth $transLength >> log_merge
+echo "done" >> log_merge
 cd ../../
 
-mv ${base}/oases/mergedAssembly/transcripts.fa ${output}
-rm -rf ${base}/oases
+#check whether the jobs completed successfully.
+for k in $Ks ; do
+   if ! [ -s ${base}/oases/directory_${k}/transcripts.fa  ] ; then
+      echo "Oases K-mer=${k} assembly failed" 
+   else
+      echo "Oases K-mer=${k} assembly succeeded" 
+      rm -rf ${base}/oases/directory_${k}
+   fi
+done
+
+if ! [ -s ${base}/oases/mergedAssembly/transcripts.fa  ] ; then
+   echo "Oases mergeAssembly failed"
+   exit 1
+else
+   echo "Oases mergeAssembly succeeded"
+   #move the output if it was a success
+   mv ${base}/oases/mergedAssembly/transcripts.fa ${output}
+   rm -rf ${base}/oases/mergedAssembly
+fi
+
